minor ref register store
